# Conflicts:
#	src/Umbraco.Core/Services/ContentEditingService.cs
#	src/Umbraco.Core/Webhooks/Events/Content/ContentEmptiedRecycleBinWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/Content/ContentMovedToRecycleBinWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/Content/ContentMovedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/Content/ContentSortedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/ContentType/DocumentTypeChangedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/ContentType/DocumentTypeDeletedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/ContentType/DocumentTypeMovedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/ContentType/DocumentTypeSavedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/ContentType/MediaTypeChangedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/ContentType/MediaTypeDeletedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/ContentType/MediaTypeMovedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/ContentType/MediaTypeSavedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/ContentType/MemberTypeChangedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/ContentType/MemberTypeDeletedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/ContentType/MemberTypeMovedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/ContentType/MemberTypeSavedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/DataType/DataTypeDeletedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/DataType/DataTypeMovedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/DataType/DataTypeSavedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/Dictionary/DictionaryItemDeletedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/Dictionary/DictionaryItemSavedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/Domain/DomainDeletedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/Domain/DomainSavedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/File/PartialViewDeletedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/File/PartialViewSavedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/File/ScriptDeletedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/File/ScriptSavedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/File/StylesheetDeletedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/File/StylesheetSavedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/File/TemplateDeletedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/File/TemplateSavedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/Language/LanguageDeletedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/Language/LanguageSavedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/Member/ExportedMemberWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/Member/MemberDeletedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/Member/MemberGroupDeletedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/Member/MemberGroupSavedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/Member/MemberSavedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/PublicAccess/PublicAccessEntryDeletedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/PublicAccess/PublicAccessEntrySavedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/Relation/RelationDeletedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/Relation/RelationSavedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/RelationType/RelationTypeDeletedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/RelationType/RelationTypeSavedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/User/AssignedUserGroupPermissionsWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/User/UserDeletedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/User/UserGroupDeletedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/User/UserGroupSavedWebhookEvent.cs
#	src/Umbraco.Core/Webhooks/Events/User/UserSavedWebhookEvent.cs
#	src/Umbraco.Infrastructure/PropertyEditors/TrueFalsePropertyEditor.cs
#	src/Umbraco.PublishedCache.HybridCache/DocumentCache.cs
#	src/Umbraco.PublishedCache.HybridCache/MediaCache.cs
#	src/Umbraco.Web.Common/Extensions/FriendlyPublishedContentExtensions.cs

+++ src/Umbraco.Web.UI.Client/src/mocks/handlers/partial-view/detail.handlers.ts
@@ -1,16 +1,26 @@
 const { rest } = window.MockServiceWorker;
+import { createProblemDetails } from '../../data/utils.js';
 import { umbPartialViewMockDB } from '../../data/partial-view/partial-view.db.js';
 import { UMB_SLUG } from './slug.js';
 import type {
-	CreateStylesheetRequestModel,
-	UpdateStylesheetRequestModel,
+	CreatePartialViewRequestModel,
+	UpdatePartialViewRequestModel,
 } from '@umbraco-cms/backoffice/external/backend-api';
 import { umbracoPath } from '@umbraco-cms/backoffice/utils';
 
 export const detailHandlers = [
 	rest.post(umbracoPath(UMB_SLUG), async (req, res, ctx) => {
-		const requestBody = (await req.json()) as CreateStylesheetRequestModel;
+		const requestBody = (await req.json()) as CreatePartialViewRequestModel;
 		if (!requestBody) return res(ctx.status(400, 'no body found'));
+
+		// Validate name
+		if (!requestBody.name) {
+			return res(
+				ctx.status(400, 'name is required'),
+				ctx.json(createProblemDetails({ title: 'Validation', detail: 'name is required' })),
+			);
+		}
+
 		const path = umbPartialViewMockDB.file.create(requestBody);
 		const encodedPath = encodeURIComponent(path);
 		return res(
@@ -25,6 +35,10 @@ export const detailHandlers = [
 	rest.get(umbracoPath(`${UMB_SLUG}/:path`), (req, res, ctx) => {
 		const path = req.params.path as string;
 		if (!path) return res(ctx.status(400));
+		if (path.endsWith('forbidden')) {
+			// Simulate a forbidden response
+			return res(ctx.status(403));
+		}
 		const response = umbPartialViewMockDB.file.read(decodeURIComponent(path));
 		return res(ctx.status(200), ctx.json(response));
 	}),
@@ -32,6 +46,10 @@ export const detailHandlers = [
 	rest.delete(umbracoPath(`${UMB_SLUG}/:path`), (req, res, ctx) => {
 		const path = req.params.path as string;
 		if (!path) return res(ctx.status(400));
+		if (path.endsWith('forbidden')) {
+			// Simulate a forbidden response
+			return res(ctx.status(403));
+		}
 		umbPartialViewMockDB.file.delete(decodeURIComponent(path));
 		return res(ctx.status(200));
 	}),
@@ -39,7 +57,11 @@ export const detailHandlers = [
 	rest.put(umbracoPath(`${UMB_SLUG}/:path`), async (req, res, ctx) => {
 		const path = req.params.path as string;
 		if (!path) return res(ctx.status(400));
-		const requestBody = (await req.json()) as UpdateStylesheetRequestModel;
+		if (path.endsWith('forbidden')) {
+			// Simulate a forbidden response
+			return res(ctx.status(403));
+		}
+		const requestBody = (await req.json()) as UpdatePartialViewRequestModel;
 		if (!requestBody) return res(ctx.status(400, 'no body found'));
 		umbPartialViewMockDB.file.update(decodeURIComponent(path), requestBody);
 		return res(ctx.status(200));
